+++ http-client/ChangeLog.md
@@ -4,6 +4,7 @@
 * Modify `instance IsString Request` to use `parseRequest` instead of `parseUrlThrow`
 * Clean up the `HttpException` constructors
 * Rename `checkStatus` to `checkResponse` and modify type
+* Fix the ugly magic constant workaround for responseTimeout
 
 ## 0.4.30